requirements.txt
albumentations==1.3.1
numpy==1.24.3
Pillow==10.2.0
PyYAML==6.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.0
